+++ sound-mind-studio/tests/test_generator_dialog.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <QObject>
+
+class GeneratorDialogTest : public QObject {
+    Q_OBJECT
+
+private slots:
+    void defaultsToLatticeFamilyAndCenteredOrderChaos();
+    void movingTheSliderChangesOrderChaos();
+    void randomizeButtonChangesTheSeed();
+};